+++ src/test/continuous/verify_alerts_spec.js
@@ -1,4 +1,7 @@
+require('../integration/integration_config.js');
+var verify_utils = require('verify_utils.js');
 var agent = require('superagent');
+var request = require('request');
 var async = require('async');
 var moment = require('moment');
 var _ = require('lodash');
@@ -6,7 +9,7 @@ var testName = __filename;
 var hubUrl = process.env.hubUrl;
 hubUrl = 'http://' + hubUrl;
 console.log(hubUrl);
-
+alertUrl = hubUrl + '/alert';
 
 /**
  * This should :
@@ -26,89 +29,12 @@ console.log(hubUrl);
 
 describe(testName, function () {
 
-    var escalationAlerts = hubUrl + '/channel/escalationAlerts';
     var verifyAlertData = hubUrl + '/channel/verifyAlertData';
-    var zomboAlertsConfig = hubUrl + '/channel/zomboAlertsConfig';
-    var zomboAlertStatus = hubUrl + '/channel/zomboAlertStatus';
-
     var existingRules = [];
 
-    it('1 - if verifyAlert rule(s) exist', function (done) {
-        agent
-            .get(zomboAlertsConfig + '/latest')
-            .accept('json')
-            .end(function (err, res) {
-                expect(err).toBe(null);
-                _.forIn(res.body.insertAlerts, function (value, key) {
-                    if (_.startsWith(key, 'verifyAlert')) {
-                        console.log('found rule', key);
-                        existingRules.push(key);
-                    }
-                });
-                done();
-            })
-    });
+    verify_utils.getExistingAlerts('verifyAlert', existingRules, hubUrl);
 
-    var escalationAlertsLinks = [];
-
-    it('gets escalationAlerts links', function (done) {
-        if (existingRules.length == 0) {
-            done();
-        } else {
-            agent
-                .get(escalationAlerts + '/time/hour')
-                .accept('json')
-                .end(function (err, res) {
-                    expect(err).toBe(null);
-                    escalationAlertsLinks = escalationAlertsLinks.concat(res.body._links.uris);
-                    agent
-                        .get(res.body._links.previous.href)
-                        .accept('json')
-                        .end(function (err, res) {
-                            expect(err).toBe(null);
-                            escalationAlertsLinks = escalationAlertsLinks.concat(res.body._links.uris);
-                            console.log('escalationAlertsLinks', escalationAlertsLinks);
-                            done();
-                        })
-
-                })
-        }
-    }, 60 * 1000);
-
-    var escalationAlertsItems = [];
-
-    it('loads escalationAlerts items', function (done) {
-        async.eachLimit(escalationAlertsLinks, 20,
-            function (link, callback) {
-                agent
-                    .get(link)
-                    .accept('json')
-                    .end(function (err, res) {
-                        expect(err).toBe(null);
-                        escalationAlertsItems.push(res.body);
-                        callback(err);
-                    })
-
-            }, function (err) {
-                console.log('escalationAlertsItems ', escalationAlertsItems.length)
-                done(err);
-            })
-    }, 60 * 1000);
-
-    it('looks for escalationAlerts items for existing rules', function (done) {
-        existingRules.forEach(function (rule) {
-            var found = 0;
-            escalationAlertsItems.forEach(function (alertItem) {
-                if (_.startsWith(alertItem.description, rule)) {
-                    console.log('found alert for rule', found, rule, alertItem);
-                    found++;
-                }
-            });
-            expect(found).toBe(1);
-        })
-
-        done();
-    });
+    verify_utils.verifyEscalationAlerts(existingRules, hubUrl);
 
     it('3 - creates verifyAlertData channel', function (done) {
         console.log('verifyAlertData', verifyAlertData);
@@ -124,35 +50,25 @@ describe(testName, function () {
 
     it('4 - create a new rule for channel verifyAlertData, named verifyAlert-{time}', function (done) {
         var name = 'verifyAlert' + moment().format('YYYY-MM-DD-HH-mm-ss');
-        agent
-            .get(zomboAlertsConfig + '/latest?stable=false')
-            .accept('json')
-            .end(function (err, res) {
-                expect(err).toBe(null);
-                //2 - delete all rules whose names start with "verifyAlert"
-                existingRules.forEach(function (name) {
-                    delete res.body.insertAlerts[name];
-                })
-                //5 - new rule be triggered for a single item on channel verifyAlertData in a minute
-                res.body.insertAlerts[name] = {
+        console.log('adding alert', name);
+        request.put({
+                url: alertUrl + '/' + name,
+                headers: {"Content-Type": "application/json"},
+                body: JSON.stringify({
                     channel: 'verifyAlertData',
                     threshold: 0,
                     serviceName: 'test',
                     operator: '>',
                     timeWindowMinutes: 1,
                     type: 'channel'
-                };
-                console.log('body', res.body.insertAlerts);
-                agent
-                    .post(zomboAlertsConfig)
-                    .type('json')
-                    .accept('json')
-                    .send(res.body)
-                    .end(function (err, res) {
-                        expect(err).toBe(null);
-                        done();
-                    })
-            })
+                })
+            },
+            function (err, response, body) {
+                expect(err).toBeNull();
+                expect(response.statusCode).toBe(201);
+                done();
+            });
+
     }, 60 * 1000);
 
     it('6 - add an item to the channel verifyAlertData', function (done) {
@@ -168,5 +84,6 @@ describe(testName, function () {
             })
     });
 
+    verify_utils.deleteExisting(existingRules);
 
 });
